Block: Adding SIO ioscheduler
SIO is based on the deadline scheduler but it's more like a mix
between no-op and deadline.
In other words, SIO is like a lighter version of deadline but it
doesn't do any kind of sorting, so it's aimed mainly for random-access
devices (like SSD hard disks) where request sorting is no needed
(as any sector can be accesed in a constant time, regardless of its
physical location).