Cpufreq: remove copy_prev_load from 'struct cpu_dbs_common_info'
because it is a bit redundant, as we also have 'prev_load' which can
store an integer value and can be used instead of 'copy_prev_load'.

True load can also turn out to be zero during long idle intervals (and
hence the actual value of 'prev_load' and the overloaded value can clash).
However this is not a problem because, if the true load was really zero
in the previous interval, it makes sense to evaluate the load afresh for
the current interval rather than copying the previous load.

So, drop 'copy_prev_load' and use 'prev_load' instead.