lib: lz4: fixed zram with lz4 on big endian machines
Based on Sergey's patch, this fixes zram with lz4 compression on big endian cpus.

Note that the 64-bit preprocessor test is not a cleanup, it's part of
the fix, since those identifiers are bogus (for example, __ppc64__
isn't defined anywhere else in the kernel, which means we'd fall into
the 32-bit definitions on ppc64).