isp: fix for potentitial array out of bound access
There is no bound check on dual_hw_ms_cmd->num_src, which is coming from userspace
num_src is used as the index for the input_src array which has a size of 5.
The current code did not check the num_src to make sure that it never exceeds the input_src array size.

Change-Id: If5927e06e70cce4afb0ae9f2cdfec80f76f83771