mmc: host: Shift pm qos unvoting to sofitrq
In eMMC driver CQ mode, the sdhc driver votes for pm qos and unvotes in
hardware irq when there are no more requests. But in a particular corner
scenario, the qos unvoting is not done by hardware irq handler after
last request and this can cause power regression. So the unvoting was
been shifted to software irq context of eMMC driver where there is
enough information to accurately know when there are no active requests to
the eMMC driver.

There is a very rare corner case when the active req counter is
checked by completion context and before it schedules the work to
unvote, the issuing context can set active req counter and do
qos voting. In this case the performance suffers but this is a
very rare corner case as the completion context is in softirq.
Also it is not good to add locking to prevent this corner case as
issuing and completion contexts should not have locking in between them.

Change-Id: I9fb714d3a5e4eb35a0bedd69691e3f1f84fc4cd8