keychord: Fix races in keychord_write.
There are multiple bugs caused by threads racing in keychord_write.
 1) Threads racing through this function can cause the same element to
be added to a linked list twice (multiple calls to
input_register_handler() for the same input_handler struct). And the
races can also cause an element in a linked list that doesn't exist
attempted to be removed (multiple calls to input_unregister_handler()
with the same input_handler struct).
 2) The races can also cause duplicate kfree's of the keychords struct.

Change-Id: I6329a4d58c665fab5d3e96ef96391e07b4941e80